Added multiple point lights and directional light

+++ src/org/lwjglb/game/engine/utils/SimplexNoise.java
@@ -43,7 +43,7 @@ public class SimplexNoise {
 	 * @param y
 	 * @return between -1 and 1
 	 */
-	public double getNoise(int x, int y) {
+	public double getNoise(double x, double y) {
 
 		double result = 0;
 
@@ -57,7 +57,7 @@ public class SimplexNoise {
 
 	}
 
-	public double getNoise(int x, int y, int z) {
+	public double getNoise(double x, double y, double z) {
 
 		double result = 0;
